Killed off the old optimize_joins file, and moved its remnants into indexes.go. Removed a bunch of unused functionality from joinColExpr
