+++ sql/analyzer/join_search_test.go
